Move RAP Background task to AnnoStore file
The code is only needed if there is an anno store. So move the
background process there. And then don't even start it if there is no
annotation server.
